requirements/test.txt
# Test dependencies go here.
-r base.txt



coverage==4.4.1
flake8==3.3.0 # pyup: != 2.6.0
pycodestyle==2.3.1
django-test-plus==1.0.17
factory-boy==2.8.1

# pytest
pytest-django==3.1.2
pytest-sugar==0.8.0
freezegun==0.3.9
Faker==0.8.9
